src/main/java/com/example/a01_Data_parsing_servers_part1/part_2/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikkel sandell",email = "mikkelsandell@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
